Refactor admin routes and components for better organization
