+++ src/components/profile/ProfileDetails.jsx
@@ -1,8 +1,8 @@
-import useForm from "../hooks/formHooks";
-import Button from "./Button";
-import Input from "./Input";
-import {useUser} from "../hooks/userHooks";
-import {useUserContext} from "../hooks/useUserContext";
+import useForm from "../../hooks/formHooks";
+import Button from "../Button";
+import Input from "../Input";
+import {useUser} from "../../hooks/userHooks";
+import {useUserContext} from "../../hooks/useUserContext";
 import {useNavigate} from "react-router-dom";
 import {useState} from "react";
