+++ src/views/Profile.jsx
@@ -3,9 +3,9 @@ import {useLanguage} from "../context/LanguageContext";
 import {useUserContext} from "../hooks/useUserContext";
 import {useUser} from "../hooks/userHooks";
 import useOrder from "../hooks/useOrder";
-import OrderTable from "../components/OrderTable";
-import ProfileImage from "../components/ProfileImage";
-import ProfileDetails from "../components/ProfileDetails";
+import OrderTable from "../components/profile/OrderTable";
+import ProfileImage from "../components/profile/ProfileImage";
+import ProfileDetails from "../components/profile/ProfileDetails";
 import LoadingWheel from "../components/LoadingWheel";
 
 const Profile = () => {
@@ -53,7 +53,7 @@ const Profile = () => {
   if (loading) {
     return <LoadingWheel />;
   }
-  console.log("Order details", orderDetails);
+
   return (
     <div>
       <article>
